+++ idlang/src/main/java/mimuw/idlearn/idlang/logic/keywords/If.java
@@ -18,7 +18,7 @@ public class If extends Expression {
 	private final Block onFalse;
 
 	public If(Expression condition, Block onTrue, Block onFalse) {
-		condition.assertType(Type.Integer);
+		condition.assertType(Type.Long);
 		this.type = Null;
 		this.condition = condition;
 		this.onTrue = onTrue;
@@ -26,7 +26,7 @@ public class If extends Expression {
 	}
 
 	public If(Expression condition, Block onTrue) {
-		condition.assertType(Type.Integer);
+		condition.assertType(Type.Long);
 		this.type = Null;
 		this.condition = condition;
 		this.onTrue = onTrue;
@@ -38,7 +38,7 @@ public class If extends Expression {
 		Value condEvaluation = condition.evaluate(scope, counter, inputScanner, outputWriter);
 		counter.addTime(delay);
 
-		if ((Integer)condEvaluation.value != 0)
+		if ((Long)condEvaluation.value != 0)
 			return onTrue.evaluate(new Scope(scope), counter, inputScanner, outputWriter);
 		if (onFalse != null)
 			return onFalse.evaluate(new Scope(scope), counter, inputScanner, outputWriter);
